PaymentForm/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from .serializer import PaymentSerializer

@api_view(['POST'])
def create_payment(request):
    serializer = PaymentSerializer(data=request.data)
    if serializer.is_valid():
        serializer.save()
        return Response(serializer.data, status=201)
    return Response(serializer.errors, status=400)
