requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
